test: move fu tests to fu module
